https://github.com/Vrushika1912/TXON_O1.git
logo design is a very basic project in graphic design. You will learn a lot in this project. Logos serve to represent a given organization or company through a visual image that can be easily understood and recognized. It is the primary source of their brand representation. Every business wants to design a unique logo. They know how these logos encourage people to take interest in the products, as well as get converted towards them.